+++ CHANGELOG.md
@@ -7,7 +7,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ## [x.x.x]
 
 ### Added
-
+- Support for CRAM alignment files
 
 ### Fixed
 - Fixed update OMIM command bug due to change in the header of the genemap2 file